ScrapBasicDriverBaseExclusive/requirements.txt
# beautifulsoup4==4.10.0
requests
beautifulsoup4
mysql-connector-python
